:hammer: refactor(jwt-authentication): UserDetail refactoring #81
